+++ cylc/flow/option_parsers.py
@@ -80,9 +80,16 @@ class OptionSettings():
             self.kwargs.update({kwarg: value})
 
     def __eq__(self, other):
-        """Args and Kwargs, but not other props equal."""
+        """Args and Kwargs, but not other props equal.
+
+        (Also make an exception for kwargs['help'] to allow lists of sources
+        prepended to 'help' to be passed through.)
+        """
         return (
-            self.kwargs == other.kwargs
+            (
+                {k: v for k, v in self.kwargs.items() if k != 'help'}
+                == {k: v for k, v in other.kwargs.items() if k != 'help'}
+            )
             and self.args == other.args
         )
 
@@ -115,6 +122,22 @@ ICP_OPTION = OptionSettings(
     metavar="CYCLE_POINT or OFFSET",
     action='store',
     dest="icp"
+)
+
+AGAINST_SOURCE_OPTION = OptionSettings(
+    ['--against-source'],
+    help=(
+        "Load the workflow configuration from the source directory it was"
+        " installed from using any options (e.g. template variables) which"
+        " have been set in the installation."
+        " This is useful if you want to see how changes made to the workflow"
+        " source would affect the installation if reinstalled."
+        " Note if this option is used the provided workflow must have been"
+        " installed by `cylc install`."
+    ),
+    dest='against_source',
+    action='store_true',
+    default=False
 )
 
 
@@ -713,9 +736,10 @@ def combine_options_pair(first_list, second_list):
                 and first & second
             ):
                 # if any of the args are different:
+
                 if first.args == second.args:
-                    # if none of the arg names are different.
-                    raise Exception(f'Clashing Options \n{first}\n{second}')
+                    raise Exception(
+                        f'Clashing Options \n{first.args}\n{second.args}')
                 else:
                     first_args = first - second
                     second.args = second - first