+++ cylc/flow/templatevars.py
@@ -22,6 +22,21 @@ from typing import Any, Dict
 from cylc.flow.exceptions import InputError
 
 
+from cylc.flow.rundb import CylcWorkflowDAO
+
+
+def get_template_vars_from_db(run_dir):
+    """Get template vars stored in a workflow run database.
+    """
+    template_vars = {}
+    if (run_dir / 'log/db').exists():
+        dao = CylcWorkflowDAO(str(run_dir / 'log/db'))
+        dao.select_workflow_template_vars(
+            lambda _, row: template_vars.__setitem__(row[0], eval_var(row[1]))
+        )
+    return template_vars
+
+
 def eval_var(var):
     """Wrap ast.literal_eval to provide more helpful error.
 
@@ -65,6 +80,7 @@ def load_template_vars(template_vars=None, template_vars_file=None):
                     continue
                 key, val = line.split("=", 1)
                 res[key.strip()] = eval_var(val.strip())
+
     if template_vars:
         for pair in template_vars:
             key, val = pair.split("=", 1)
@@ -82,6 +98,4 @@ def get_template_vars(options: Values) -> Dict[str, Any]:
     Returns:
         template_vars: Template variables to give to a Cylc config.
     """
-    return load_template_vars(
-        options.templatevars, options.templatevars_file
-    )
+    return load_template_vars(options.templatevars, options.templatevars_file)