+++ tests/unit/test_templatevars.py
@@ -15,13 +15,19 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import pytest
+import sqlite3
 import tempfile
 import unittest
 
 from types import SimpleNamespace
 
+
 from cylc.flow.exceptions import PluginError
-from cylc.flow.templatevars import get_template_vars, load_template_vars
+from cylc.flow.templatevars import (
+    get_template_vars_from_db,
+    get_template_vars,
+    load_template_vars
+)
 
 
 class TestTemplatevars(unittest.TestCase):
@@ -102,3 +108,48 @@ class TestTemplatevars(unittest.TestCase):
             'none': None
         }
         self.assertEqual(expected, load_template_vars(template_vars=pairs))
+
+
+@pytest.fixture(scope='module')
+def _setup_db(tmp_path_factory):
+    tmp_path = tmp_path_factory.mktemp('test_get_old_tvars')
+    logfolder = tmp_path / "log/"
+    logfolder.mkdir()
+    db_path = logfolder / 'db'
+    conn = sqlite3.connect(db_path)
+    conn.execute(
+        r'''
+            CREATE TABLE workflow_template_vars (
+                key,
+                value
+            )
+        '''
+    )
+    conn.execute(
+        r'''
+            INSERT INTO workflow_template_vars
+            VALUES
+                ("FOO", "42"),
+                ("BAR", "'hello world'"),
+                ("BAZ", "'foo', 'bar', 48"),
+                ("QUX", "['foo', 'bar', 21]")
+        '''
+    )
+    conn.commit()
+    conn.close()
+    yield get_template_vars_from_db(tmp_path)
+
+
+@pytest.mark.parametrize(
+    'key, expect',
+    (
+        ('FOO', 42),
+        ('BAR', 'hello world'),
+        ('BAZ', ('foo', 'bar', 48)),
+        ('QUX', ['foo', 'bar', 21])
+    )
+)
+def test_get_old_tvars(key, expect, _setup_db):
+    """It can extract a variety of items from a workflow database.
+    """
+    assert _setup_db[key] == expect